Move Entry conditions
